+++ src/calendar/calendar.controller.ts
@@ -5,16 +5,13 @@ import {
   Get,
   Param,
   Post,
-  Put,
   UseGuards,
 } from '@nestjs/common';
 import { CalendarService } from './calendar.service';
-import { GroupEventDto } from './dto/groupEvent.dto';
-import { UserEventDto } from './dto/userEvent.dto';
+import { GroupEventDto } from './dto/event.group.dto';
+import { UserEventDto } from './dto/event.user.dto';
 import { JwtAuthGuard } from 'src/auth/guards/jwt.auth.guard';
 import { GetUser } from 'src/common/decorator/get.user.decorator';
-import { UpdateUserEventDto } from './dto/updateUserEvent.dto';
-import { UpdateGroupEventDto } from './dto/updategroupEvent.dto';
 
 @Controller('/api/calendar')
 export class CalendarController {
